Admin Course Management section enabled
New course creation feature with rich text editor available on dashboard.

+++ src/ie/ucc/team19/dao/SqlMapper.java
@@ -9,7 +9,7 @@ public class SqlMapper {
         //javaToSqlMap.put("java.lang.String", java.sql.Types.CHAR);
         javaToSqlMap.put("java.math.BigDecimal", java.sql.Types.NUMERIC);
         javaToSqlMap.put("java.math.BigDecimal", java.sql.Types.DECIMAL);
-        javaToSqlMap.put("java.lang.Boolean", java.sql.Types.INTEGER);
+        javaToSqlMap.put("java.lang.Boolean", java.sql.Types.TINYINT);
         javaToSqlMap.put("java.lang.Byte", java.sql.Types.TINYINT);
         javaToSqlMap.put("java.lang.Short", java.sql.Types.SMALLINT);
         javaToSqlMap.put("java.lang.Integer", java.sql.Types.INTEGER);
